Upkeep & User info in Reviews
Removed unnecessary code in campus.html to instead extend and be in block content. Also gave text a black outline since it is hard to read on the image; still needs better styling.

Removed index.html and renamed index2 html to home so we don't have a weirdly named file.

When reviewing, will now display username or Anonymous if not logged in.

Also changed various spacing.